refactor: Name 클래스를 생성해 분리하고 View 의 inputCarNames가 Name 리스트를 반환하도록 변경
