+++ src/main/java/controller/Race.java
@@ -2,13 +2,13 @@ package controller;
 
 import java.util.List;
 
-import domain.Car;
 import domain.MoveCount;
+import domain.Name;
 import view.InputView;
 
 public class Race {
 	public void startRace() {
-		List<Car> cars = InputView.inputCarNames();
+		List<Name> names = InputView.inputCarNames();
 		MoveCount moveCount = InputView.inputMoveCount();
 	}
 }
